Revert "prefilter app and droplet queries when spaces are provided"
The solution to this works, but is broken if apps per space is very
large. Reverting this revert might be a good idea in the future, but
its not easy to evaluate when the tradeoff is worthwhile, since doing
this correctly with a join will forego composite key speedup.

It might be better to categorize this as "premature optimization."

This reverts commit 6ae99f3598ff9e6329919a65e435d5c9513354b0.