Allow updating service instance with inactive plan
Previously, non-admin users could not update the service instance(name, tags)
if the plan is not active/public anymore. With this fix, they can.
Parameters are still not updateable since might they require an update to the instance,
which might not be allowed by the broker, since the plan is disabled.
Admins can still update all attributes.

[finishes #163189599]
[PR]

Co-authored-by: Aarti Kriplani <akriplani@pivotal.io>
Signed-off-by: Aarti Kriplani <akriplani@pivotal.io>