Add path and mode support to v3 resource matches
We want the response from POST /v3/resource_matches to be passed to the
v3 package upload endpoint, so that clients can take advantage of the
resource pool caching. The package upload endpoint expects to receive
the file path and the file mode for each resource, which we are adding
in this commit.

A future commit will allow the package upload endpoint to be able to
receive this new v3-style format for a resource (i.e. fn => path, sha1
=> checksum, etc)

[#164610893]

Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>
Co-authored-by: Mike Kenyon <mkenyon@pivotal.io>