Begin refactoring error handling in AppApplyManifestActionJob
This change comes from a team discussion a few months back regarding
the "errsanity" and large stack of error classes being rescued in
the AppApplyManifestActionJob. We have this stack of errors explicitly
specified because they are known validation errors with user-suitable
error messages that come from each sub-action that the manifest executes.

We decided to move the exception handling of the sub-actions closer to where
they actually occur to bubble up a single AppApplyManifest::InvalidManifest
error. This commit begins this effort by making this change for the
ProcessScale::InvalidProcess error.

[#164903447]

Authored-by: Tim Downey <tdowney@pivotal.io>