Split ServiceEventRepository into two classes:
- ServiceEventRepository::WithBrokerActor
- ServiceEventRepository::WithUserActor

This is done because there are two distinct kind of service events being
reported, and one kind requires a user_audit_info object, and another
one doesn't. To be able to only instantiate a single kind (e.g.
WithBrokerActor) splitting those up is beneficial.

To not make this a crazy big refactoring touching all the code in the
system, we've decided for older code to keep calling
ServiceEventRepository directly, and that will delegate to
with_user_actor or with_broker_actor.

[#165969720]

Co-authored-by: George Blue <gblue@pivotal.io>