make annotation keys conform to metadata key prefix rules
ADR: decisions/0004-adding-key-prefix-to-annotations.md

one awkward thing at the end of this is that we're not sure we can
ever remove or rename the legacy "key" column from the annotations table. The ADR
says we should write a time bomb test, but it's unclear if 6 months will
suffice as "enough time" to be renaming columns. If you read this in the
distant future, know that we thought about trying to rename key_name to key
to remove the alias.

[finishes #166447964]

Co-authored-by: Connor Braa <cbraa@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Chris Selzo <cselzo@pivotal.io>
Co-authored-by: Seth Boyles <sboyles@pivotal.io>