Use fully qualified column names in DB query
In #1384 a user described an error that occurs when querying service plans with a
service broker filter. The error were:

- PG::AmbiguousColumn: ERROR: column reference "id" is ambiguous
- PG::AmbiguousColumn: ERROR: column reference "active" is ambiguous

The captured DB query was:
```
SELECT count(*) AS "count" FROM "service_plans"
  LEFT JOIN "services" ON ("services"."id" = "service_plans"."service_id")
  LEFT JOIN "service_brokers" ON ("service_brokers"."id" = "services"."service_broker_id")
  WHERE ((("public" IS TRUE) OR ("id" IN (<long list of IDs>)))
    AND ("active" IS TRUE) AND ("service_brokers"."guid" = '<guid>'))
  LIMIT 1
```

It can be seen that the `active` and `id` columns are not qualified by
table name, so the DB is confused when doing a join where multiple
tables may contain the `active` and `id` columns.

The solution is to qualify the column names with the table name.

[finishes #166974930]
[PR]

Co-authored-by: George Blue <gblue@pivotal.io>