Reduce flakiness of failed job assertions in app_manifests_spec.rb
Certain environments (such as Travis CI) would fail with errors like
the following in this file:

```
Failure/Error: api_errors = YAML.safe_load(job.cf_api_error)['errors']

     TypeError:
       no implicit conversion of nil into String
```

It was difficult to reproduce this locally, but my theory is that lazily
loading the job in the `let` could be more likely to load the wrong job.

Moved the job fetching closer to where they are worked off and added
additional assertions to verify we're using the correct one.

Authored-by: Tim Downey <tdowney@pivotal.io>