Add support for filtering by null values
Before this change, there was no way in V3 to filter by `null` values. For
example, you could use `/v3/buildpacks?stacks=cflinuxfs2` to get all
buildpacks with `stack: cflinuxfs2`, but no way to get those with
`stack: null`. Now, `/v3/buildpacks?stacks=` (no value) will return
those buildpacks.

There are two list endpoints with query parameters where this makes a
difference:

- `/v3/buildpacks?stacks=` returns buildpacks with no stack
- `/v3/apps?stacks=` returns apps using a buildpack with no stack

Note that the meaning of an "empty filter" (`?fields=`) is now
resource-specific. In most places, it means empty string (`""`), but in
these two places it now means `null`.

Going forward, when a fetcher needs to support filtering by `null` for a
query parameter, it can use the `NullFilterQueryGenerator` (introduced
in this change). This handles adding the correct `WHERE` clause to the
query to include `null` when needed.

[Finishes #167150073]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>