v3: Add `status.value` and `status.reason` to deployments
Before this change, deployment state was exposed only as the `state`
field on the deployment resource. This field was somewhat misleading and
hard for clients to use (e.g. `state: DEPLOYED` could either mean the
deployment completed successfully, or it was superseded by another
deployment for the same app, so it didn't actually complete).

This change introduces two new fields on deployments that, together,
make it easier for clients to understand the current "state" of a
deployment:

- `status.value` can be `DEPLOYING` (still in-progress, CC is still
trying to complete it) or `FINALIZED` (CC is done trying to complete it,
whether it succeeded or not)

- `status.reason` gives more information about the status. Currently,
possible reasons are `DEPLOYED` (deployment completed successfully),
`CANCELED` (deployment was canceled by a user before completion), or
`SUPERSEDED` (a newer deployment was created for the same app, so CC
will now start working towards that one).

The `state` field still exists, but the recommendation now is for
clients to not use it to determine deployment state. Eventually, it will
be removed (no longer exposed to clients).

[Finishes #167205520]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Sannidhi Jalukar <sjalukar@pivotal.io>