v3 docs: preserve location hash on redirect
The docs automatically redirect from /index.html to
/version/${LATEST_VERSION}/index.html. But if you clicked a link to
/index.html#apps, without this change you would be redirected to
/version/.../index.html instead of /version/.../index.html#apps.

Note that this is a best-effort attempt to preserve the hash. If, for
example, the user has JavaScript disabled in their browser (so the
automatic JS-based redirect wouldn't work), they will get redirected by
the <meta http-equiv="refresh"> tag to /version/.../index.html without
the hash (at least in Chrome, this is the behavior). If this doesn't
work, like if the user has automatic refreshes disabled in their
browser, they could click the "Click here if you are not redirected"
link and end up at /version/.../index.html, also without the hash. So
the fallback case is just what we had before this change.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>