Reset process ports to default when unmapping last
Previously, when unmapping the last route from a app/process we would
remove the last port in the `ports` array (even the default 8080). This
causes process with health_check_type set to `port` (like the `web`
process) to fail validations (`ports array cannot be empty`).

This commit makes it when the last route is being unmapped, we will set
the `ports` field back to nil, allowing the system to fallback to the
default (8080 for buildpack, -1 for docker).

[Finishes #167477477]

Co-authored-by: Luan Santos <lsantos@pivotal.io>
Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>