🐞 Correctly map manifest-defined, non-8080 destinations to Docker apps
Broken behavior: Docker apps would not be mapped to the appropriate port
when using the apply-manifest endpoints. Regardless on which port the
Docker container listened, cc would always create destinations pointing
to port 8080. Users would see a 404 when attempting to access their
container app when the Docker container listened to a different port;
e.g. the `nginx:latest` Docker image listens on port 80.

fixes: <https://github.com/cloudfoundry/capi-release/issues/149>

[finishes #167518505]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>