Don't create V3 package for mis-matched lifecycle-type
If a user attempts to upload a Docker container to a buildpack app (or
vice-versa) we should see a helpful error message rather than allow it
to be created and fail (in a rather cryptic manner) during staging (see
below for cryptic output).

We elected *not* to move the controller tests for `/v3/packages` into
request specs; we wanted the scope of this commit to be as tight as
possible in the event it needed to be reverted.

Similarly, we elected *not* to backport this to the V2 endpoint, for we
weren't sure of the effect it would have on V2; see
[Slack](https://cloudfoundry.slack.com/archives/CBNDLQBB5/p1564096532007800)
for more context (the conversation is not threaded).

fixes (in the CLI, when trying to stage a bits package for a Docker app):
```
   Staging...
   missing flag: dockerRef required
     -cacheDockerImage
       	Caches Docker images to private docker registry
     -dockerDaemonExecutablePath string
       	path to the 'docker' executable (default "/tmp/docker_app_lifecycle/docker")
     (...snip...)
   Exit status 1
   Staging Failed: STG: Exited with status 1
```

Related Issues:
- [Creating a package with docker metadata for an application with a `buildpack` lifecycle does not error](https://github.com/cloudfoundry/cloud_controller_ng/issues/1262)
- [Docker app shouldn't be able to have a bits package](https://github.com/cloudfoundry/cloud_controller_ng/issues/1311)

[fixes #1311]
[fixes #1262]
[finishes #167461778]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>