Add audit events to space delete/update
- Deleting a space now creates an audit event.
- Updating a space now creates an audit event.
- Audit events are also created for the recursive deletions, e.g. app
deletion
- Added ORG_MANAGER to the list of roles that can update a space. This
change led to adding org_guid to the list of args for
queryer.can_update_space? which now takes space_guid and org_guid.
Previously, only space_guid was required as an argument since the list
of roles that can update a space used to only be SPACE_MANAGER. This
change propogated in many small changes to reflect the additional arg
required.
- Removed the delete space audit event creation from the V2 spaces
controller since this is now being done in the DeleteSpace action which
is shared between V2 and V3.

[finishes #167742979]

Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>
Co-authored-by: Brendan Smith <brsmith@pivotal.io>
Co-authored-by: Harsha Nandiwada <hnandiwada@pivotal.io>