Allow API user creation with POST /v3/users
- for now this endpoint has been implemented for non-client users.
- presentation_name is a new field for /v3/users. It represents the
display name for users. This was introduced in V3 because non-user
clients do not have usernames -- they have Client IDs that will be used
as the presentation_name while the username field will be returned as
null.
- username and presentation_name fields are equal for non-client users.
For user creation when the users are not registered in UAA, these fields
will be returned as null.
- User creation was also added to V3 Docs. `Users` is now an
experimental resource with this new create endpoint documented.

[finishes #164176715]

Co-authored-by: Piyali Banerjee <pbanerjee@pivotal.io>
Co-authored-by: Harsha Nandiwada <hnandiwada@pivotal.io>
Co-authored-by: Luan Santos <lsantos@pivotal.io>