Allow API Client to delete a user
- includes removal of org and space roles for deleted user
- includes doc changes
- Rather than the 404 originally specified in the story, global auditors
will receive a 403. This is because in general, only users with
write-permissions are allowed to pursue a delete action. Global
auditors, just like read-only admins, only have read-permissions and
therefore the beforeaction checking for write-permissions for delete
endpoints fail before getting to the actual controller-level method.
Global auditors are not authorized to perform a delete action.

[finishes #164176792]

Co-authored-by: Supraja Narasimhan <snarasimhan@pivotal.io>