Role join tables include GUID, updated/create times
We expand the roles tables to include GUIDs (because, hey, we're using
GUIDs everywhere else even though the primary key, a
monotonically-increasing integer, was perfectly adequate) and two
timestamps: `created_at` and `updated_at`. Note that every entry in the
join table has a unique GUID--it's not unique by user or by organization
or by space, it's unique by join. The same user can be a space_auditor
in several spaces, and each join would have a unique GUID.

We have two migrations:

- the first migration creates the new columns & defaults the GUID to
null and the timestamps to the current time.

- the 2nd migration populates the GUIDs and updates the `updated_at`
timestamp.

The 2nd migration is particulary important; it must finish quickly. In
this case, it runs in 23 seconds on a 500k table. Across 7 tables, this
works out to under 3 minutes, which is under our ten-minute budget.
Earlier versions of this algorithm took 3 hours.

CAVEAT:

If roles are created using the v2 endpoint while this migration is
happening, they will not get GUIDs (null). To address this, we plan to
add a hook on the `/v3/roles` list endpoint to fill in missing GUIDs if
a null GUID is detected.

GUID AMBIGUITY

Adding the `guid` column introduced ambiguity in portions of the code
where `guid` wasn't explicitly associated with a table. We addressed
this by explicitly defining the table (e.g. `users__guid`).

fixes:
```
     Sequel::DatabaseError:
       PG::AmbiguousColumn: ERROR:  column reference "guid" is ambiguous
       LINE 1: SELECT "guid" FROM "users" INNER JOIN "spaces_developers" ON...
```

[#164176815]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>