only create start usage events for RUNNING tasks
- App Usage Events[0] are used to inform consumers (such as billing
systems) of changes of app/task state that affect resource consumption
on the underlying platform
- A CF Task is not actually consuming resources until it is scheduled on
the runtime and transitions into the RUNNING state so this change
creates the `TASK_STARTED` event when the task is RUNNING instead of
PENDING
- A separate bug (https://www.pivotaltracker.com/story/show/152831505)
where the Task Syncer ignores PENDING tasks forever can cause tasks that
are never scheduled to linger in the PENDING state for a very long time.
This change mitigates the impact of that issue on billing.

Related issue: https://github.com/cloudfoundry/cloud_controller_ng/issues/1365

[0] https://docs.cloudfoundry.org/running/managing-cf/usage-events.html#app-usage-events

Authored-by: Tim Downey <tdowney@pivotal.io>