🐞 PATCH Environment Variable Groups accepts {}
`PATCH /v3/environment_variable_groups/:name` accepts {} (empty hash) as
valid input (previously it would error with "must be a hash, Var must be
a hash").

We now pass an optional flag (`populate_empty_hash_with_empty_var`) to
the `UpdateEnvironmentVariablesMessage` constructor which directs it to
include a required key. The constructor is shared, so we chose not to
change its default behavior.

- re-organized `request/environment_variable_groups_spec.rb`; it has the
  original tests plus and additional one for the empty hash `{}` case
- corrected two typos & re-formatted

[#167953997]

Authored-by: Brian Cunnie <bcunnie@pivotal.io>