v3: Allow role creation by username and origin
Previously, POST /v3/roles only supported creating roles for users by
passing the user's guid. This change adds support for specifying the
user by username (and origin). When there's only one user with the given
username, the origin is optional. But when only a username is provided
and the same username exists in multiple origins, we return an error.

[Finishes #164343021]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>