🐞 `PATCH /v3/environment_variable_groups` errors
Prior behavior: when patching with a non-object `var` key, it would emit
two redundant errors:

```
"detail": "must be an object, Var must be an object"
```

We addressed this by creating a new validator incorporating the behavior
of the two earlier validators (the source of the redundant error
messages were an overlap in the validators). This enabled us to remove
one of the validators (it was no longer needed) while leaving the other
validator unchanged (to avoid inadvertantly modifying V2 behavior).

[finishes #167953997]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>