Merge service broker states into service_brokers table
- Previously the states were in a separate service_broker_states table
because states are a V3 concept that V2 is unaware of
- Because there was always a 1:1 mapping, it makes more sense to store
the state in the service_brokers table, and this has no impact on V2
- This also fixes a bug where old service broker states were not being
cleaned up (see https://www.pivotaltracker.com/story/show/169595982)
- Because the service_broker_states table is dropped, this change will
lose data about service brokers created in V3 that are in an unavailable
state. The V3 service-brokers endpoint is still experimental, and any
users of experimental endpoints will expect this kind of change.

[finishes #169595982]
[PR]

Co-authored-by: Derik Evangelista <devangelista@pivotal.io>