v3: Simplify Role model query by flattening unions
The Role model is backed by a large union of 7 different tables
(representing the 7 different roles). The way we had written this,
Sequel was generating a nested SQL query that looked like:

```
SELECT
  *
FROM
  (
    SELECT
      *
    FROM
      (
        SELECT
          *
        FROM
          (
            SELECT
              *
            FROM
              (
                SELECT
                  *
                FROM
                  (
                    SELECT
                      *
                    FROM
                      (
                        SELECT
                          'organization_user' AS "type",
                          "role_guid" AS "guid",
                          "user_id",
                          "organization_id",
                          -1 AS "space_id",
                          "created_at",
                          "updated_at"
                        FROM
                          "organizations_users"
                        UNION
                        ...
```

This change does two things to improve this query:

The first is to add `all: true`, which changes the `UNION`s to `UNION ALL`s.
We know that these rows should be unique, so this should be fine and will maybe
give us better performance.

The second is to add `from_self: false` to each of the unions, and then
add a `.from_self` at the very end of the whole dataset. By default,
Sequel nests each additional union in a subquery. This gives the deeply
nested structure shown above. By only doing a single `from_self` at the
end, and not at each union along the way, we keep the SQL flat and close
to what we might write by hand:

```
SELECT
  *
FROM
  (
    SELECT
      'organization_user' AS "type",
      "role_guid" AS "guid",
      "user_id",
      "organization_id",
      -1 AS "space_id",
      "created_at",
      "updated_at"
    FROM
      "organizations_users"
    UNION ALL
     (
        SELECT
          'organization_manager' AS "type",
    ...
```

Authored-by: Reid Mitchell <rmitchell@pivotal.io>