v3: Fix existence check query for role-deletion jobs
The v3 roles resource is not backed by a table. It's a union of 7
individual roles tables. When we implemented the DELETE /v3/roles
endpoint, we allowed jobs to be created with `resource_type` of `role`
(as expected). These jobs would execute correctly, but when we tried to
fetch the job by guid, we'd get a 500 back.

The underlying error was the following, caused by the query made in
PollableJobModel:

```
Mysql2::Error: Table 'cloud_controller.roles' doesn't exist
```

In this query, we were inspecting the job's `resource_type`, pluralizing
it, and making a query to the corresponding table. In the case of roles,
there is no table called `roles`, so we were failing.

The simplest possible fix was just to special-case roles: for roles, we
go through the Role model, which is backed by the union dataset.

[Finishes #164176828]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>