client sees a **💎 luxurious error 💎** when attempting to create a **space-level role**
When creating a space-level role, the user will see the
more-informative, "Users cannot be assigned roles in a space if they do
not have a role in that space's organization." instead of the original,
"Invalid user. Ensure that the user exists and you have access to it."
(though that message is still in use when creating an _org-level_ role).

To help support
[#169440492](https://www.pivotaltracker.com/story/show/169440492), we
want all users able to create space roles to see the more luxurious
error.

[Finishes #169688314]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>