v3: Fix error message when creating role for non-existent user by name
Previously, the error was:

```
No user exists with the username 'does-not-exist'.
```

Now, for space roles, the error will be:

```
Users cannot be assigned roles in a space if they do not have a role in that space's organization.
```

For org roles, the error is unchanged.

Also, changed a bunch of assertions like
`expect(last_response.status).to eq(422)` to use the `have_status_code`
matcher, which gives us better messages when it fails.

[Finishes #169688314]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>