**API** client **filters** users by **username and origin**
Typical request:
```
GET /v3/users?usernames=username1,username2&origins=my-origin
```
returns users with username `username1` or `username2` in origin `my-origin`

When filtering on origins, usernames _must_ be included in the filter.
We are constrained by UAA's design—UAA only allows filtering by origins
when usernames are present.

When multiple usernames are specified, the users returned are the union
of the list of usernames. Similarly with origins.

When both usernames and origins are specified, only the usernames that
are present in the origins specified are returned (it's the
intersection, not the union, of the usernames and origins).

[#169737392]

Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>