🐞 `POST /v3/spaces/:guid/actions/apply_manifest`
Check that the name of the app, when routes are not provided, is
DNS-compliant (i.e. only a-z, A-Z, 0-9, _, -).

For example, POSTing with the following manifest:

```yaml
applications:
- name: "%%%%"
```

Would result in an error message similar to the following:

```
"%%%" must contain only alphanumeric characters, "_", or "-" when routes are not present
```

Bugs:

- We don't check that the length 63 characters or less

[#168811666]

Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>