🐞 `GET /v3/users?usernames=:names` w/out UAA
Prior (broken) behavior: When UAA is unavailable _and_ a query of users
is done which hits the UAA endpoint, the resulting error was
uninformative ("... unknown error ..."):

```json
{
   "errors": [
      {
         "title": "UnknownError",
         "detail": "An unknown error occurred.",
         "code": 10001
      }
   ]
}
```

The new error is more informative (UAA is not available):

```json
{
   "errors": [
      {
         "detail": "The UAA service is currently unavailable",
         "title": "CF-UaaUnavailable",
         "code": 20004
      }
   ]
}
```

We accomplish this by rescuing `CF::UAA::BadTarget`, raising
`VCAP::CloudController::UaaEndpointDisabled`, rescuing that in turn, and
raising a final error with an informative message.

Other Changes
-----

- We check the HTTP status using `have_status_code()` instead of `eq()`;
`have_status_code()` provides much richer error output when the tests
fail, and make troubleshooting easier.

- We reformatted one of the spec files; the change (a lambda function)
is whitespace-only.

[finishes #169737392]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>