Remove UAA query in role delete job
A previous change added a UAA query to the role deletion job in order to
access the username for the user whose role was being deleted. This job
is asynchronously executed by a cloud controller worker, but the worker
instances are not configured with the necessary properties to execute
this UAA query. This caused the role deletion to eventually fail for an
"unknown" error.

This commit moves the query up to the controller level so that it can be
executed by the API instance, which has the required configuration. Role
deletion should now work as expected.

[#170059486](https://www.pivotaltracker.com/story/show/170059486)

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>