Allow role creation for users not in CC database
A role can now be requested for a user who does not exist in the cloud
controller database if the user exists in the UAA database. In this
scenario, a new User object will be created with the username or user
guid in the request, and the requested role will be assigned to that
user.

This change in behavior only effectively applies to the creation of org
roles. For space roles, if a user were to be created, the role creation
would fail because the user would not have any organization-level
roles. The newly created user would then have to be immediately rolled
back, and the error message would be identical to not finding a valid
user.
Because of this, the implementation only has code changes in the creation of
org roles to keep the code as simple as possible. We've included an
integration test to reflect this behavior.

[#170061485](https://www.pivotaltracker.com/story/show/170061485)

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>