`POST /v2/quota_definitions` → unlimited memory_limit
We introduce the option to set an "unlimited" memory_limit with the
sentinel value `-1`

```json
{"memory_limit":-1}
```

The "unlimited" feature is a requirement (and default) of V3, so we're
backporting to V2, for it entails a change to the `quota_definition`
model; to whit, a negative number (-1) is now allowed for
`memory_limit`.

- We have changed the `less_than_zero` error to match other validations
for fields that allow -1, i.e. `invalid_memory_limit`.

- Updated docs to include -1 as the unlimted value for POST and PUT

[#164973386]

Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>