Set maximum value validation for app instances
Previously there were no maximum value validations for app instances.
This was causing a 500 CF-Database error if a user tried to push an app
with instances higher than the max value for an INT type in
My/PostgresSQL. Now there is a maximum value validation for instances
in the ManifestProcessScaleMessage set to 2,000,000 and the 500 no
longer occurs.

[Finishes #166848802]

Authored-by: Piyali Banerjee <pbanerjee@pivotal.io>