v3: Fix resource type for org quota delete jobs
By default, DeleteActionJobs get a `resource_type` generated
automatically based on the model name. This means that org quotas were
getting a `resource_type` of `quota_definition` (the name of the
model/DB table), and the corresponding jobs were getting an `operation`
field of `quota_definition.delete`.

In v3, we want these jobs to have `operation:
organization_quota.delete`, so with this commit we are subclassing
`DeleteActionJob` for this specific case and overriding the
`resource_type` method to return the new v3 type.

Also, a couple of minor docs tweaks for org quotas.

[Finishes #164973539]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>