Change the metadata representation
At the moment, an API is returning a string field:

```
{
  "metadata": "\n{\"test\": \"yes\"}"
}
```

This means the client would need to parse the string into an object in
order to use it. It's not practical and provides an overhead.

We can remove the JSON parsing which will cause the object to expand
further into:

```
{
  "metadata": {"test": "yes"}
}
```