v3: Fix race condition on service broker deletes
We are reordering a few lines in `service_brokers_controller`. We
noticed that when we updated the state of the service broker to
`DELETE_IN_PROGRESS` after we enqueued the async job, we sometimes got
500s with this underlying error:

```
Attempt to update object did not result in a single row modification (SQL: UPDATE `service_brokers` SET `updated_at` = CURRENT_TIMESTAMP,
 `state` = 'DELETE_IN_PROGRESS' WHERE (`id` = 667) LIMIT 1)
```

...because the row had been deleted before the update.

This change makes sure we update the state on the broker before
enqueuing the job.

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Weyman Fung <wfung@pivotal.io>