v3: Restore security group listing functionality
There were some commits that were reverted/applied a few times over, and
some of this code got lost along the way. The result was that the
controller was not even using the new fetcher, so the endpoint was not
working (and we also lost the tests for it). This commit brings back the
missing code, including tests and docs.

[Finishes #168655161]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>