Ensure `BUILDPACK_SET` usage event is recorded for v3 staging
In v2, we always recorded a `BUILDPACK_SET` app usage event after an app
was successfully staged and started (since these were combined into the
same workflow).

In v3, staging was split out into a separate step from starting (it
happens when a build is created). But we lost the `BUILDPACK_SET` event
in this codepath. This was causing a CATS test to fail when run with the
v7 CF CLI, which uses v3 endpoints to create a build and initiate
staging.

This change ensures that even if `with_start` is not sent (and it is not
set in the v3 codepaths), we record this event for each process in the
app when staging completes.

[Finishes #171634008]

Authored-by: Reid Mitchell <rmitchell@pivotal.io>