Ensure kpack build pods labeled with source_type=STG
As noted in https://github.com/cloudfoundry/cf-k8s-logging/issues/18,
staging logs do not show up when pushing an app in cf-for-k8s.

The solution to this is to label each build pod with
`cloudfoundry.org/source_type=STG`, which the logging infrastructure
will pick up and set on the log envelopes.  The CLI _only_ displays logs
with the `STG` source_type when pushing an app.

[#171991349]

Authored-by: Chris Selzo <cselzo@pivotal.io>