V3: `POST /v3/routes` on TCP domains: no host/path
TCP domains don't accept `host` or `path` in the body of the request.
`path` is an HTTP attribute, and thus makes no sense for a TCP domain.
Although you could make the argument that `host` is not an HTTP
attribute, its purpose is to enable [Name-based Virtual Host
Support](https://httpd.apache.org/docs/2.4/vhosts/name-based.html),
which again is HTTP-thing. This is usually accomplished with a wildcard
DNS entry (e.g. `*.domain.tld).

Many Cloud Foundry installations have a TCP domain, with an IP
address/load balancer distinct from the shared app domain. Often the
DNS entry's hostname is `tcp`, as in `tcp.domain.tld`.

The convention we are using to identify a TCP domain is the following:
- if it's running on k8s, it's an "http" domain (this will change)
- if `router_group.guid` is populated, it's a "tcp" domain.

- The route controller has an awkward method,
  `unprocessable_non_http_protocols(message, domain)`; its sole purpose is
  to appease RuboCop's `Metrics/CyclomaticComplexity` requirements. Yes,
  we are not happy about it either.

- We yanked the `protocols` method (which returns the domain's protocols,
  currently "http" and "tcp") out of the presenter and thrust it into the
  model, which is a more appropriate home. We also backfilled tests for
  it.

[finishes #169636853]

Co-authored-by: Mona Mohebbi <mmohebbi@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>