V3: `GET /v3/usage_events/:guid`
Usage events are a synthesis of app usage events and service usage
events. We combine both `AppUsageEvent` and `ServiceUsageEvent` in a
model `UsageEvent` backed by a union of the two tables.

Usage Events and their models do **not** have an `updated_at` field, for
events are created but not updated; however, we present an `updated_at`
field to the user for completeness's sake. The `updated_at` is the same
as the `created_at` value.

[#172985853]

[finishes #169237831]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>