V3: client can **see data** for a **service** usage event
We present the `service_guid` and `service_label` columns from the
`service_usage_event` table as `service_offering.guid` and
`service_offering.name` to reflect the new naming conventions across V3.

Unlike the story, we do not display the organization name, it is not in
the model (neither the app usage event model nor the service usage event
model), and Greg has
[agreed](https://cloudfoundry.slack.com/archives/CCRLX88DD/p1590701929419700)
to not display it (V2 does not display the org name, either).

[finishes #172985880]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>