v3: Run usage event-purging specs in isolation: truncation mode
When we added these tests recently, we started seeing lots of random
other failures in CI when run against MySQL. It seemed to be a test
pollution issue (other tests were seeing DB entries they didn't expect,
or were missing DB entries that they did expect).

After talking to the CAPI team, we learned that the table truncation
performed by these new endpoints may change the mode that tests are run
in. Usually, they are run in a transaction that gets rolled back
automatically, so we avoid pollution. But this truncation may change
that.

To fix it, we are attempting here to run these new truncating tests with
`isolation: :truncation`, to keep them better isolated and avoid test
pollution.

Authored-by: Reid Mitchell <rmitchell@pivotal.io>